[FS-55] - refactor: transform drawer into component
