+++ src/pages/dashboard/components/sidebar/components/drawer-header/index.tsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/material/styles";
+
+export const DrawerHeader = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "flex-end",
+  padding: theme.spacing(0, 1),
+  // necessary for content to be below app bar
+  ...theme.mixins.toolbar,
+}));
