+++ src/pages/dashboard/components/sidebar/sidebar.tsx
@@ -5,9 +5,7 @@ import ConnectWithoutContactIcon from "@mui/icons-material/ConnectWithoutContact
 import _isEqual from "lodash/isEqual";
 import cryptoJs from "crypto-js";
 
-import { styled, Theme, CSSObject } from "@mui/material/styles";
 import Divider from "@mui/material/Divider";
-import MuiDrawer from "@mui/material/Drawer";
 
 import { TypeMenuItem } from "constants/type-menu-item";
 
@@ -24,53 +22,6 @@ import { TUser } from "models/user";
 import { healthcheckMqtt } from "services/resources/engine-mqtt";
 import { useSelector } from "react-redux";
 import { RootState } from "store";
-
-const drawerWidth = 240;
-
-const openedMixin = (theme: Theme): CSSObject => ({
-  width: drawerWidth,
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: "hidden",
-});
-
-const closedMixin = (theme: Theme): CSSObject => ({
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: "hidden",
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up("sm")]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "flex-end",
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: "nowrap",
-  boxSizing: "border-box",
-  ...(open && {
-    ...openedMixin(theme),
-    "& .MuiDrawer-paper": openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    "& .MuiDrawer-paper": closedMixin(theme),
-  }),
-}));
 
 type Props = {
   isOpen: boolean;
@@ -228,9 +179,11 @@ export const Sidebar: React.FC<Props> = ({ isOpen }) => {
 
   return (
     <>
-      <Drawer variant="permanent" open={isOpen}>
-        <DrawerHeader />
+      <S.Drawer open={isOpen}>
+        <S.DrawerHeader />
+
         <Divider />
+
         <S.CustomList>
           <div>
             {sidebar.menuItems.map((menuItem, index) =>
@@ -262,7 +215,7 @@ export const Sidebar: React.FC<Props> = ({ isOpen }) => {
             <S.Text>{version}</S.Text>
           </S.VersionContainer>
         </S.CustomList>
-      </Drawer>
+      </S.Drawer>
 
       {sidebar.isOpenDialog && (
         <ProcessIdSearch
